+++ src/app/(auth)/login/page.tsx
@@ -5,8 +5,8 @@ import { useRouter } from 'next/navigation';
 import Link from 'next/link';
 import { useMutation } from '@tanstack/react-query';
 import { useAuth } from '@/providers/auth-provider'; // Zustand 대신 새 Provider 사용
-import SuccessModal from '@/components/common/SuccessModal';
-import FailModal from '@/components/common/FailModal';
+import SuccessModal from '@/app/components/SuccessModal';
+import FailModal from '@/app/components/FailModal';
 import { PATHS } from '@/types/auth';
 
 // JWT 로그인 API 함수
