+++ src/app/(auth)/register/page.tsx
@@ -4,8 +4,8 @@ import { useState } from 'react';
 import { useRouter } from 'next/navigation';
 import Link from 'next/link';
 import { useMutation } from '@tanstack/react-query';
-import SuccessModal from '@/components/common/SuccessModal';
-import FailModal from '@/components/common/FailModal';
+import SuccessModal from '@/app/components/SuccessModal';
+import FailModal from '@/app/components/FailModal';
 
 const registerUser = async (userData: {
   username: string;
