--- src/components/layout/Banner/index.tsx
@@ -1,13 +1,0 @@
-export default function Banner() {
-  return (
-    <div className="bg-blue-600 text-white">
-      <div className="container mx-auto px-4 py-4">
-        <div className="max-w-3xl mx-auto text-center">
-          <h2 className="text-lg font-bold">
-            개발자 스터디 모임에 오신 것을 환영합니다
-          </h2>
-        </div>
-      </div>
-    </div>
-  );
-}
